requirements.txt
fastapi>=0.115.4
uvicorn>=0.32.0
pydantic>=2.10.4
python-dotenv>=1.0.1
llm>=0.19
llm-gemini>=0.10
aiofiles>=23.2.1
